started using map structs for setup. see description
need to fix make error that is caused by including a header in a header. need to actually pass a struct into every function and not just the variables inside the struct. this is because in future ill be passing a linkedlist of structs around that can be undone, peeked at, etc