big problems with freeing, see description of commit
tried a bunch of different ways. i think the underlying issue is that each node is pointing to the same map pointers. to resolve: i need to make a function that creates a new char** for each node but also transfers the current characters on the current map to the new map. just loop through and add to the new